moved test/ out of src/
