+++ test/runTest.ts
@@ -1,5 +1,4 @@
 import * as path from 'path';
-
 import { runTests } from 'vscode-test';
 
 async function main(): Promise<void> {
