+++ test/suite/groups.test.ts
@@ -1,61 +1,60 @@
 import * as assert from 'assert';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { softwareRegex } from '../../helpers';
-import { configSection, extensionID, ignoreConsoleLogs, resetState, setTestConfig } from '../suite/testHelpers';
+import { groupRegex } from '../../src/helpers';
+import { configSection, extensionID, ignoreConsoleLogs, resetState, setTestConfig } from './testHelpers';
 
 
-describe('Software', function () {
+describe('Groups', function () {
     const testPath: string = path.resolve(__dirname, '..', '..', '..', 'src', 'test', 'files', 'test.md');
     const testUri: vscode.Uri = vscode.Uri.file(testPath);
     let ext: vscode.Extension<unknown> | undefined;
-    let modifiedConfig: vscode.WorkspaceConfiguration;
+    let modifiedConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration(configSection);
 
     before(async function () {
         ignoreConsoleLogs();
         ext = vscode.extensions.getExtension(extensionID);
         await ext?.activate();
         exports = ext?.exports;
-        modifiedConfig = vscode.workspace.getConfiguration(configSection);
-        await setTestConfig('software', true, modifiedConfig);
+        await setTestConfig('groups', true, modifiedConfig);
     });
     after(async function () {
-        await setTestConfig('software', undefined, modifiedConfig);
+        await setTestConfig('groups', undefined, modifiedConfig);
     });
     beforeEach(ignoreConsoleLogs);
     afterEach(resetState);
 
-    it('regex should match software ID', async function () {
-        const gid = 'S0045';
-        assert.ok(softwareRegex.test(gid));
+    it('regex should match group ID', async function () {
+        const gid = 'G0007';
+        assert.ok(groupRegex.test(gid));
     });
-    it('should provide one competion item for exact software IDs', async function () {
-        const expectedSID = 'S0045';
-        const position: vscode.Position = new vscode.Position(11, expectedSID.length);
+    it('should provide one competion item for exact group IDs', async function () {
+        const expectedGID = 'G0007';
+        const position: vscode.Position = new vscode.Position(9, expectedGID.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
         assert.ok(results.items[0] instanceof vscode.CompletionItem);
-        assert.strictEqual(results.items[0].label, expectedSID);
+        assert.strictEqual(results.items[0].label, expectedGID);
         assert.strictEqual(results.items[0].kind, vscode.CompletionItemKind.Value);
     });
-    it('should provide all completion items containing a software name', async function () {
-        const expectedName = 'ADVSTORESHELL';
-        const position: vscode.Position = new vscode.Position(12, expectedName.length);
+    it('should provide all completion items containing a group name', async function () {
+        const expectedName = 'APT28';
+        const position: vscode.Position = new vscode.Position(10, expectedName.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
-        const softwareResults: Array<vscode.CompletionItem> = results.items.filter((item: vscode.CompletionItem) => {
-            return item instanceof vscode.CompletionItem && softwareRegex.test(`${item.detail}`);
+        const groupResults: Array<vscode.CompletionItem> = results.items.filter((item: vscode.CompletionItem) => {
+            return item instanceof vscode.CompletionItem && groupRegex.test(`${item.detail}`);
         });
-        assert.strictEqual(softwareResults.length, 1);
-        assert.ok(softwareResults[0] instanceof vscode.CompletionItem);
-        assert.strictEqual(softwareResults[0].label, expectedName);
-        assert.strictEqual(softwareResults[0].kind, vscode.CompletionItemKind.Value);
+        assert.strictEqual(groupResults.length, 1);
+        assert.ok(groupResults[0] instanceof vscode.CompletionItem);
+        assert.strictEqual(groupResults[0].label, expectedName);
+        assert.strictEqual(groupResults[0].kind, vscode.CompletionItemKind.Value);
     });
-    it('should provide a hover for software IDs', async function () {
-        const expectedSID = 'S0045';
-        const expectedLink = `[Source Link](https://attack.mitre.org/software/${expectedSID})`;
-        const position: vscode.Position = new vscode.Position(11, expectedSID.length);
+    it('should provide a hover for group IDs', async function () {
+        const expectedGID = 'G0007';
+        const expectedLink = `[Source Link](https://attack.mitre.org/groups/${expectedGID})`;
+        const position: vscode.Position = new vscode.Position(9, expectedGID.length);
         const expectedRange: vscode.Range = new vscode.Range(new vscode.Position(position.line, 0), new vscode.Position(position.line, position.character));
         const results = await vscode.commands.executeCommand('vscode.executeHoverProvider', testUri, position);
         assert.ok(results instanceof Array);
@@ -67,33 +66,32 @@ describe('Software', function () {
     });
 });
 
-describe('Software Settings', function () {
+describe('Group Settings', function () {
     // bumping timeout on this due to config updates in afterEach()
     // ... potentially taking a long time
     this.timeout(5000);
     const testPath: string = path.resolve(__dirname, '..', '..', '..', 'src', 'test', 'files', 'test.md');
     const testUri: vscode.Uri = vscode.Uri.file(testPath);
-    const modifiedConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration(configSection);;
+    const modifiedConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration(configSection);
 
     before(async function () {
-        await setTestConfig('software', true, modifiedConfig);
+        await setTestConfig('groups', true, modifiedConfig);
     });
-    beforeEach(function () {
+    beforeEach(async function () {
         ignoreConsoleLogs();
     });
     afterEach(async function () {
         resetState();
     });
     after(async function () {
-        await setTestConfig('software', undefined, modifiedConfig);
+        await setTestConfig('groups', undefined, modifiedConfig);
         await setTestConfig('completionFormat', undefined, modifiedConfig);
     });
     it('completionFormat: should show only an ID when set to id', async function () {
-        const sid = 'S0045';
-        const expectedDetail: string = sid;
-        await setTestConfig('software', true, modifiedConfig);
+        const gid = 'G0007';
+        const expectedDetail: string = gid;
         await setTestConfig('completionFormat', 'id', modifiedConfig);
-        const position: vscode.Position = new vscode.Position(11, sid.length);
+        const position: vscode.Position = new vscode.Position(9, gid.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
@@ -101,11 +99,10 @@ describe('Software Settings', function () {
         assert.strictEqual(results.items[0].detail, expectedDetail);
     });
     it('completionFormat: should show only a name when set to name', async function () {
-        const sid = 'S0045';
-        const expectedDetail = 'ADVSTORESHELL';
-        await setTestConfig('software', true, modifiedConfig);
+        const gid = 'G0007';
+        const expectedDetail = 'APT28';
         await setTestConfig('completionFormat', 'name', modifiedConfig);
-        const position: vscode.Position = new vscode.Position(11, sid.length);
+        const position: vscode.Position = new vscode.Position(9, gid.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
@@ -113,11 +110,10 @@ describe('Software Settings', function () {
         assert.strictEqual(results.items[0].detail, expectedDetail);
     });
     it('completionFormat: should show only a link when set to link', async function () {
-        const sid = 'S0045';
-        const expectedDetail = 'https://attack.mitre.org/software/S0045';
-        await setTestConfig('software', true, modifiedConfig);
+        const gid = 'G0007';
+        const expectedDetail = 'https://attack.mitre.org/groups/G0007';
         await setTestConfig('completionFormat', 'link', modifiedConfig);
-        const position: vscode.Position = new vscode.Position(11, sid.length);
+        const position: vscode.Position = new vscode.Position(9, gid.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
@@ -125,11 +121,10 @@ describe('Software Settings', function () {
         assert.strictEqual(results.items[0].detail, expectedDetail);
     });
     it('completionFormat: should show only a name when set to fullname', async function () {
-        const sid = 'S0045';
-        const expectedDetail = 'ADVSTORESHELL';
-        await setTestConfig('software', true, modifiedConfig);
+        const gid = 'G0007';
+        const expectedDetail = 'APT28';
         await setTestConfig('completionFormat', 'fullname', modifiedConfig);
-        const position: vscode.Position = new vscode.Position(11, sid.length);
+        const position: vscode.Position = new vscode.Position(9, gid.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
@@ -137,11 +132,10 @@ describe('Software Settings', function () {
         assert.strictEqual(results.items[0].detail, expectedDetail);
     });
     it('completionFormat: should show an ID and name when set to id-name', async function () {
-        const sid = 'S0045';
-        const expectedDetail = `${sid} ADVSTORESHELL`;
-        await setTestConfig('software', true, modifiedConfig);
+        const gid = 'G0007';
+        const expectedDetail = `${gid} APT28`;
         await setTestConfig('completionFormat', 'id-name', modifiedConfig);
-        const position: vscode.Position = new vscode.Position(11, sid.length);
+        const position: vscode.Position = new vscode.Position(9, gid.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
@@ -149,29 +143,27 @@ describe('Software Settings', function () {
         assert.strictEqual(results.items[0].detail, expectedDetail);
     });
     it('completionFormat: should show an ID and name when set to id-fullname', async function () {
-        const sid = 'S0045';
-        const expectedDetail = `${sid} ADVSTORESHELL`;
-        await setTestConfig('software', true, modifiedConfig);
+        const gid = 'G0007';
+        const expectedDetail = `${gid} APT28`;
         await setTestConfig('completionFormat', 'id-fullname', modifiedConfig);
-        const position: vscode.Position = new vscode.Position(11, sid.length);
+        const position: vscode.Position = new vscode.Position(9, gid.length);
         const results = await vscode.commands.executeCommand('vscode.executeCompletionItemProvider', testUri, position);
         assert.ok(results instanceof vscode.CompletionList);
         assert.strictEqual(results.items.length, 1);
         assert.ok(results.items[0] instanceof vscode.CompletionItem);
         assert.strictEqual(results.items[0].detail, expectedDetail);
     });
-    it('should enable the Software providers when set to true', async function () {
-        await setTestConfig('software', true, modifiedConfig);
-        const expectedSID = 'S0045';
-        const position: vscode.Position = new vscode.Position(11, expectedSID.length);
+    it('should enable the Group providers when set to true', async function () {
+        const expectedGID = 'G0007';
+        const position: vscode.Position = new vscode.Position(9, expectedGID.length);
         const results = await vscode.commands.executeCommand('vscode.executeHoverProvider', testUri, position);
         assert.ok(results instanceof Array);
         assert.strictEqual(results.length, 1);
     });
-    it('should disable the Software providers when set to false', async function () {
-        await setTestConfig('software', false, modifiedConfig);
-        const expectedSID = 'S0045';
-        const position: vscode.Position = new vscode.Position(11, expectedSID.length);
+    it('should disable the Group providers when set to false', async function () {
+        await setTestConfig('groups', false, modifiedConfig);
+        const expectedGID = 'G0007';
+        const position: vscode.Position = new vscode.Position(9, expectedGID.length);
         const results = await vscode.commands.executeCommand('vscode.executeHoverProvider', testUri, position);
         assert.ok(results instanceof Array);
         assert.strictEqual(results.length, 0);
